+++ contracts/fund_test.go
@@ -102,13 +102,11 @@ func TestPerformAccountFunding(t *testing.T) {
 		ID:                 types.FileContractID{1},
 		HostKey:            hk1,
 		RemainingAllowance: types.Siacoins(1),
-		Good:               true,
 	})
 	store.contracts = append(store.contracts, Contract{
 		ID:                 types.FileContractID{2},
 		HostKey:            hk1,
 		RemainingAllowance: types.Siacoins(2),
-		Good:               true,
 	})
 
 	// add h2 with one contract
@@ -121,7 +119,6 @@ func TestPerformAccountFunding(t *testing.T) {
 		ID:                 types.FileContractID{3},
 		HostKey:            hk2,
 		RemainingAllowance: types.Siacoins(1),
-		Good:               true,
 	})
 
 	// add h3, which is unusable
@@ -131,7 +128,6 @@ func TestPerformAccountFunding(t *testing.T) {
 		ID:                 types.FileContractID{4},
 		HostKey:            hk3,
 		RemainingAllowance: types.Siacoins(1),
-		Good:               true,
 	})
 
 	// add h4, which is blocked
@@ -141,7 +137,6 @@ func TestPerformAccountFunding(t *testing.T) {
 		ID:                 types.FileContractID{5},
 		HostKey:            hk4,
 		RemainingAllowance: types.Siacoins(1),
-		Good:               true,
 	})
 
 	// fund accounts
@@ -152,7 +147,7 @@ func TestPerformAccountFunding(t *testing.T) {
 
 	// assert there were two calls, one for each usable host
 	if len(amMock.calls) != 2 {
-		t.Fatal("unexpected", len(amMock.calls))
+		t.Fatal("unexpected")
 	}
 	call1 := amMock.calls[0]
 	call2 := amMock.calls[1]
