+++ contracts/renew.go
@@ -53,16 +53,17 @@ func (cm *ContractManager) renewContract(ctx context.Context, contract Contract,
 		hc, err := cm.dialer.DialHost(ctx, host.PublicKey, host.SiamuxAddr())
 		if err != nil {
 			contractLog.Debug("failed to dial host", zap.Error(err))
-			return fmt.Errorf("failed to dial host: %w", err)
+			return nil
 		}
 		defer hc.Close()
 		res, err := hc.RenewContract(ctx, host.Settings, contract.ID, proofHeight)
 		if err != nil {
 			contractLog.Debug("failed to renew", zap.Error(err))
-			return fmt.Errorf("failed to renew contract: %w", err)
+			return nil
 		}
 		renewed := res.Contract
 		minerFee := res.RenewalSet.Transactions[len(res.RenewalSet.Transactions)-1].MinerFee
+
 		if err := cm.store.AddRenewedContract(ctx, contract.ID, renewed.ID, renewed.Revision, host.Settings.Prices.ContractPrice, minerFee, types.ZeroCurrency); err != nil {
 			return fmt.Errorf("failed to store renewed contract: %w", err)
 		}
