+++ persist/postgres/slabs.go
@@ -6,8 +6,6 @@ import (
 	"errors"
 	"fmt"
 
-	"github.com/jackc/pgx/v5"
-	proto "go.sia.tech/core/rhp/v4"
 	"go.sia.tech/core/types"
 	"go.sia.tech/indexd/slabs"
 )
@@ -58,79 +56,6 @@ func (s *Store) Slab(ctx context.Context, slabID slabs.SlabID) (slab slabs.Slab,
 		return rows.Err()
 	})
 	return
-}
-
-// Slabs returns the slabs with the given IDs from the database.
-func (s *Store) Slabs(ctx context.Context, accountID proto.Account, slabIDs []slabs.SlabID) ([]slabs.Slab, error) {
-	if len(slabIDs) == 0 {
-		return nil, nil
-	}
-
-	results := make([]slabs.Slab, len(slabIDs))
-	err := s.transaction(ctx, func(ctx context.Context, tx *txn) error {
-		dbIDMap := make(map[int64]int)
-		var dbIDs []int64
-		slabBatch := &pgx.Batch{}
-		for i, slabID := range slabIDs {
-			slabBatch.Queue(`SELECT s.id, s.encryption_key, s.min_shards, s.pinned_at
-				FROM slabs s
-				INNER JOIN account_slabs ac ON s.id = ac.slab_id
-				INNER JOIN accounts a ON a.id = ac.account_id
-				WHERE digest = $1 AND a.public_key = $2`, sqlHash256(slabID), sqlPublicKey(accountID)).QueryRow(func(row pgx.Row) error {
-				results[i].ID = slabID
-				var dbID int64
-				if err := row.Scan(&dbID, (*sqlHash256)(&results[i].EncryptionKey), &results[i].MinShards, &results[i].PinnedAt); err != nil {
-					if errors.Is(err, sql.ErrNoRows) {
-						err = slabs.ErrSlabNotFound
-					}
-					return fmt.Errorf("failed to get slab %q: %w", slabID, err)
-				}
-				dbIDs = append(dbIDs, dbID)
-				dbIDMap[dbID] = i
-				return nil
-			})
-		}
-		if err := tx.Tx.SendBatch(ctx, slabBatch).Close(); err != nil {
-			return fmt.Errorf("failed to get slabs: %w", err)
-		}
-
-		sectorsBatch := &pgx.Batch{}
-		for _, slabID := range dbIDs {
-			sectorsBatch.Queue(`SELECT s.sector_root, h.public_key, csm.contract_id
-FROM sectors s
-INNER JOIN slab_sectors ss ON s.id = ss.sector_id
-LEFT JOIN hosts h ON h.id = s.host_id
-LEFT JOIN contract_sectors_map csm ON s.contract_sectors_map_id = csm.id
-WHERE ss.slab_id = $1
-ORDER BY ss.slab_index ASC`, slabID).Query(func(rows pgx.Rows) error {
-				defer rows.Close()
-				for rows.Next() {
-					var sector slabs.Sector
-					var hostKey sql.Null[sqlPublicKey]
-					var contractID sql.Null[sqlHash256]
-
-					if err := rows.Scan((*sqlHash256)(&sector.Root), &hostKey, &contractID); err != nil {
-						return fmt.Errorf("failed to scan sector: %w", err)
-					}
-
-					if hostKey.Valid {
-						sector.HostKey = (*types.PublicKey)(&hostKey.V)
-					}
-					if contractID.Valid {
-						sector.ContractID = (*types.FileContractID)(&contractID.V)
-					}
-					results[dbIDMap[slabID]].Sectors = append(results[dbIDMap[slabID]].Sectors, sector)
-				}
-				return rows.Err()
-			})
-		}
-		if err := tx.Tx.SendBatch(ctx, sectorsBatch).Close(); err != nil {
-			return fmt.Errorf("failed to get slab sectors: %w", err)
-		}
-
-		return nil
-	})
-	return results, err
 }
 
 // PinnedSlab retrieves a pinned slab from the database by its ID.
